Add FormTypes.js
